// fem-vision-studio/design-mockup/src/lib/types/concrete/Overlappingssituatie.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * Komen er overlappingslassen in de kolom voor? Bepaalt A_s,max volgens de NB
 * bij §9.5.2(3).
 */
export type Overlappingssituatie = "GeenLassen" | "LassenBuitenDezeDoorsnede" | "TerPlaatseVanLas";